requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.11.2
bleach==6.0.0
braceexpand==0.1.7
build==0.10.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091799509/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
docker-pycreds==0.4.0
docutils==0.19
einops==0.6.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
gdown==4.6.4
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.12.1
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1655369107642/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
jaraco.classes==3.2.3
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
jeepney==0.8.0
joblib==1.2.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1633454794268/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1678994160714/work
keyring==23.13.1
kiwisolver==1.4.4
kornia==0.6.10
laion-clap==1.1.4
lazy_loader==0.1
librosa==0.10.0
lightning-utilities==0.7.1
llvmlite==0.39.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
more-itertools==9.1.0
msgpack==1.0.4
multidict==6.0.4
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1673482170163/work
pandas==1.5.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathtools==0.1.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==3.1.0
pooch==1.7.0
progressbar==2.5
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
protobuf==4.22.0
psutil==5.9.4
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1672682006896/work
pyparsing==3.0.9
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytorch-lightning==1.9.4
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==19.0.2
readme-renderer==37.3
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
resampy==0.4.2
rfc3986==2.0.0
rich==13.3.4
rsa==4.9
SciencePlots==2.0.1
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.9.1
seaborn==0.12.2
SecretStorage==3.3.3
sentry-sdk==1.16.0
setproctitle==1.3.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
soundfile==0.12.1
soupsieve==2.4
soxr==0.3.4
ssr-eval==0.0.6
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
taming-transformers==0.0.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.2.28
timm==0.6.12
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchlibrosa==0.1.0
torchmetrics==0.11.3
torchvision==0.14.1+cu116
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827245914/work
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
transformers==4.26.1
twine==4.0.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
urllib3==1.26.14
vector-quantize-pytorch==1.2.2
wandb==0.13.10
Wave==0.0.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
webdataset==0.2.35
webencodings==0.5.1
Werkzeug==2.2.3
wget==3.2
yarl==1.8.2
zipp==3.15.0
